Include koinos .proto files and fix token.hpp bug

+++ docker/scripts/build.sh
@@ -22,7 +22,9 @@ if [ $? -eq 0 ]; then
    cd src
    cp contract.wasm /src
    cd ../types
-   cp types.pb /src
+   if [ -e types.pb ]; then
+      cp types.pb /src
+   fi
 fi
 
 # cleanup
